Fix for scenario:
User opens a .dax file, attempts to edit the file, intellisense code is invoked, an Object Reference exception is thrown due to the lack of a connection, and Intellisense is disabled for the entire window.  Instead, use the already existing null check to simply not attempt to show the Intellisense window, and when a connection is made later, Intellisense will work.